pyproject.toml
[tool.poetry]
name = "overmind_core"
version = "0.1.0"
description = "Core observability backend for distributed tracing"
authors = ["Akhat", "Tyler"]
readme = "README.md"
packages = [{include = "overmind_core"}]

[tool.poetry.dependencies]
python = "^3.12"
openai = ">=1.0.0"
pydantic = {extras = ["email"], version = "^2.11.9"}
pydantic-settings = ">=2.8.1"
requests = ">=2.31.0"
tqdm = ">=4.65.0"
fastapi = ">=0.104.0"
uvicorn = ">=0.23.2"
sqlalchemy = ">=2.0.22"
alembic = ">=1.12.0"
httpx = ">=0.24.0"
cryptography = ">=41.0.0"
bcrypt = ">=4.0.1"
asyncpg = "^0.30.0"
python-jose = "^3.5.0"
python-multipart = "^0.0.22"
protobuf = "^6.33.5"
opentelemetry-proto = "^1.35.0"
opentelemetry-api = "^1.36.0"
opentelemetry-sdk = "^1.36.0"
opentelemetry-exporter-otlp-proto-http = "^1.36.0"
toolz = "^1.0.0"
cachetools = "^6.1.0"
pymupdf = "^1.26.4"
valkey-glide = ">=2.1.1"
pillow = "^12.1.1"
pyjwt = "^2.11.0"
litellm = {extras = ["google"], version = "^1.81.6"}
json-repair = "^0.56.0"
tenacity = ">=9.0.0"
celery = "5.6.2"
flower = "^2.0.1"
faker = "^40.4.0"
valkey = "^6.1.1"
redis = "^7.1.1"
async-lru = "^2.1.0"
boto3 = {version = "^1.40.10", optional = true}

[tool.poetry.extras]
aws = ["boto3"]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
pytest-cov = ">=4.1.0"
ruff = "^0.12.0"
pre-commit = "^4.5.1"
python-dotenv = "^1.2.1"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
exclude = [
    "alembic/versions/*",
]
